+++ game/platforms/openGL/gameGraphicsGL.cpp
@@ -674,7 +674,7 @@ void startCountingSpritePixelsDrawn() {
 
 
 
-unsigned int endCountingSpritePixelsDrawn() {
+double  endCountingSpritePixelsDrawn() {
     return SpriteGL::endCountingPixelsDrawn();
     }
 
@@ -683,7 +683,7 @@ unsigned int endCountingSpritePixelsDrawn() {
 // more efficient and simpler to always count
 // instead of keeping an "are we counting" state
 // just set this to zero whenever user asks to start counting
-static unsigned int numSpritesDrawn = 0;
+static double numSpritesDrawn = 0;
 
 
 void startCountingSpritesDrawn() {
@@ -692,7 +692,7 @@ void startCountingSpritesDrawn() {
 
 
 
-unsigned int endCountingSpritesDrawn() {
+double endCountingSpritesDrawn() {
     return numSpritesDrawn;
     }
